+++ src/core/enums/index.ts
@@ -1,0 +1,1 @@
+export * from './storage-type.enum';
