+++ src/core/enums/storage-type.enum.ts
@@ -1,0 +1,4 @@
+export enum StorageType {
+    S3 = 'S3',
+    Mongo = 'MONGO',
+}
