+++ src/core/interfaces/db-config.interface.ts
