+++ src/core/interfaces/env.interface.ts
@@ -4,4 +4,6 @@ export interface IEnv {
     DB_DATABASE: string;
     DB_USER: string;
     DB_PASSWORD: string;
+    QUOTES_BUCKET_NAME: string;
+    STORAGE_TYPE: string;
 }
