+++ src/core/interfaces/health-check.interface.ts
