+++ src/core/interfaces/index.ts
@@ -1,5 +1,4 @@
 export * from './health-check.interface';
-export * from './quote.interface';
 export * from './db-config.interface';
 export * from './mapper.interface';
 export * from './env.interface';
