+++ src/core/interfaces/mapper.interface.ts
