+++ src/core/services/health-check.service.ts
