+++ src/core/services/index.ts
@@ -1,0 +1,2 @@
+export * from './health-check.service';
+export * from './storage.service';
