+++ src/quotes/interfaces/quote.interface.ts
