+++ src/quotes/models/quote.model.ts
