+++ src/quotes/routers/index.ts
@@ -1,0 +1,1 @@
+export * from './quote.router';
