+++ src/quotes/services/index.ts
@@ -1,0 +1,2 @@
+export * from './quote-storage.service';
+export * from './quote.service';
